Make PatchLoadBalancerServicesStep an abstract class
Make PatchLoadBalancerServicesStep an abstract class to allow easy
overridable of services / models. This is needed because traefik-rgw is
conditional.

sunbeam.jobs should never import anything from commands, both a as
design choice and to prevent import loops. Extract k8s helper to the
jobs package.
