+++ sunbeam-python/sunbeam/jobs/k8s.py
@@ -1,0 +1,105 @@
+# Copyright (c) 2024 Canonical Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+# implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import ipaddress
+import logging
+
+from snaphelpers import Snap
+
+from sunbeam.jobs.common import SunbeamException
+
+LOG = logging.getLogger(__name__)
+
+# --- Microk8s specific
+MICROK8S_CLOUD = "sunbeam-microk8s"
+MICROK8S_KUBECONFIG_KEY = "Microk8sConfig"
+MICROK8S_DEFAULT_STORAGECLASS = "microk8s-hostpath"
+METALLB_ANNOTATION = "metallb.universe.tf/loadBalancerIPs"
+
+# --- K8s specific
+K8S_CLOUD = "sunbeam-k8s"
+K8S_DEFAULT_STORAGECLASS = "csi-rawfile-default"
+K8S_KUBECONFIG_KEY = "K8SKubeConfig"
+SERVICE_LB_ANNOTATION = "io.cilium/lb-ipam-ips"
+
+SUPPORTED_K8S_PROVIDERS = ["k8s", "microk8s"]
+
+
+def validate_cidr_or_ip_range(ip_ranges: str):
+    for ip_range in ip_ranges.split(","):
+        ips = ip_range.split("-")
+        if len(ips) == 1:
+            if "/" not in ips[0]:
+                raise ValueError(
+                    "Invalid CIDR definition, must be in the form 'ip/mask'"
+                )
+            ipaddress.ip_network(ips[0])
+        elif len(ips) == 2:
+            ipaddress.ip_address(ips[0])
+            ipaddress.ip_address(ips[1])
+        else:
+            raise ValueError(
+                "Invalid IP range, must be in the form of 'ip-ip' or 'cidr'"
+            )
+
+
+class K8SHelper:
+    """K8S Helper that provides cloud constants."""
+
+    @classmethod
+    def get_provider(cls) -> str:
+        """Return k8s provider from snap settings."""
+        provider = Snap().config.get("k8s.provider")
+        if provider not in SUPPORTED_K8S_PROVIDERS:
+            raise SunbeamException(
+                f"k8s provider should be one of {SUPPORTED_K8S_PROVIDERS}"
+            )
+
+        return provider
+
+    @classmethod
+    def get_cloud(cls) -> str:
+        """Return cloud name matching provider."""
+        match cls.get_provider():
+            case "k8s":
+                return K8S_CLOUD
+            case _:
+                return MICROK8S_CLOUD
+
+    @classmethod
+    def get_default_storageclass(cls) -> str:
+        """Return storageclass matching provider."""
+        match cls.get_provider():
+            case "k8s":
+                return K8S_DEFAULT_STORAGECLASS
+            case _:
+                return MICROK8S_DEFAULT_STORAGECLASS
+
+    @classmethod
+    def get_kubeconfig_key(cls) -> str:
+        """Return kubeconfig key matching provider."""
+        match cls.get_provider():
+            case "k8s":
+                return K8S_KUBECONFIG_KEY
+            case _:
+                return MICROK8S_KUBECONFIG_KEY
+
+    @classmethod
+    def get_loadbalancer_annotation(cls) -> str:
+        """Return loadbalancer annotation matching provider."""
+        match cls.get_provider():
+            case "k8s":
+                return SERVICE_LB_ANNOTATION
+            case _:
+                return METALLB_ANNOTATION
